pyproject.toml
[project]
name = "POC-AI"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "camilo_bpt", email = ""},
]
dependencies = ["databricks-sql-connector>=4.0.0", "openai>=1.66.3", "shiny>=1.3.0", "python-dotenv>=1.0.1", "pyarrow>=19.0.1", "fastapi>=0.115.11"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
